Former-commit-id: a58081fc063449a8e47a416b2e826fb24621a87e

+++ cpp/src/db/engine/ExecutionEngineImpl.h
@@ -18,7 +18,7 @@
 #pragma once
 
 #include "ExecutionEngine.h"
-#include "src/wrapper/vec_index.h"
+#include "src/wrapper/VecIndex.h"
 
 #include <memory>
 #include <string>
@@ -101,7 +101,7 @@ protected:
     std::string location_;
 
     int32_t nlist_ = 0;
-    int64_t gpu_num_ = 0;
+    int32_t gpu_num_ = 0;
 };
